Created SearchBarComponent and ThemeSelectionComponent
